Split monolithic file, add configurable export path, and audit logging
- Break the single ~470-line .cs file into one class per file
  (Script, AssemblyInfo, AnalysisResult, PortalDoseSnapshot,
  GammaCalculator, MainView, ActivityLogger) and update the .csproj
  Compile items accordingly
- Resolve the leftover template TODOs: move assembly version attributes
  into AssemblyInfo.cs and declare [ESAPIScript(IsWriteable = false)]
  explicitly, since the tool only reads portal dose images and writes
  a CSV to local disk
- Replace the hardcoded C:\Temp\PortalDosimetryReports export path with
  a SaveFileDialog (still defaulting to that folder when writable), so
  export no longer fails silently when the user lacks permissions there
- Add ActivityLogger: a minimal, best-effort audit trail (timestamp,
  user, patient/plan/field, pass/fail/error counts, exported file path)
  for institutional traceability of QA runs
